+++ src/pages/Home/Home.jsx
@@ -1,24 +1,75 @@
 import Styles from './Home.module.css'
 import burger from '../../assets/img/Rectangle 313.png'
-import P1 from '../../assets/img/Rectangle 314.png'
-import P2 from '../../assets/img/Rectangle 315.png'
-import P3 from '../../assets/img/Rectangle 316.png'
-import P4 from '../../assets/img/Rectangle 317.png'
-import P5 from '../../assets/img/Rectangle 318.png'
-import P6 from '../../assets/img/Rectangle 319.png'
 import Footer from '../../components/footer/Footer'
 import 'react-multi-carousel/lib/styles.css';
 import Section1 from './section/Section1'
-import Section2 from './section/Section2'
+import Carousel from 'react-multi-carousel';
+import three from '../../assets/img/Make Vector.png'
+import Product from './section/Product'
+import React, { useEffect, useState } from 'react'
+import axios from 'axios'
+
 
 const Home = () => {
- 
+  const responsive = {
+    superLargeDesktop: {
+      // the naming can be any, depends on you.
+      breakpoint: { max: 4000, min: 1024 },
+      items: 2,
+      slidesToSlide: 2
+    },
+    desktop: {
+      breakpoint: { max: 1024, min: 800 },
+      items: 2,
+      slidesToSlide: 2
+    },
+    tablet: {
+      breakpoint: { max: 800, min: 464 },
+      items: 2,
+      slidesToSlide: 2
+    },
+    mobile: {
+      breakpoint: { max: 464, min: 0 },
+      items: 1
+    }
+  };
+
+  const [recipes, setRecipes] = useState([])
+  useEffect(() => {
+    axios.get('https://food-recipe-be.onrender.com/recipes')
+      .then((res) => {
+        console.log(res.data.data);
+        setRecipes(res.data.data);
+      })
+  }, []);
+
+  const products = recipes.map(recipe => (
+    <Product title={recipe.title} photo={recipe.photo} />
+  ))
 
   return (
     <>
       <div className={Styles.bdy}>
         <Section1 />
-        <Section2 />        
+        <section className={Styles.se}>
+          <div className="container mt-5">
+            <div className="row">
+              <div className={Styles.slice}></div>
+              <h5 className={Styles.hd5}>Popular For you !</h5>
+            </div>
+          </div>
+          <div className="mt-5">
+            <img src={three} className={Styles.n} alt="" />
+          </div>
+          <div className={Styles.wrap}>
+            <Carousel 
+            responsive={responsive}>
+              {products}
+            </Carousel>
+          </div>
+
+        </section >
+
         <section className={Styles.sec}>
           <div className="container mt-5">
             <div className="row">
@@ -58,42 +109,14 @@ const Home = () => {
           </div>
           <div className="container mt-5">
             <div className='row'>
-              <div className="col-md-4">
+              {recipes.map((recipe=>(
+                <div className="col-md-4">
                 <div className={Styles.wrappe1}>
-                  <img src={P1} alt="" className={Styles.product} />
-                  <h6 className={Styles.titles}>Chiken Kare</h6>
+                  <img src={recipe.photo} alt="" className={Styles.product} />
+                  <h6 className={Styles.titles}>{recipe.title}</h6>
                 </div>
               </div>
-              <div className="col-md-4">
-                <div className={Styles.wrappe2}>
-                  <img src={P2} alt="" className={Styles.product} />
-                  <h6 className={Styles.titles}>Bomb Chicken</h6>
-                </div>
-              </div>
-              <div className="col-md-4">
-                <div className={Styles.wrappe3}>
-                  <img src={P3} alt="" className={Styles.product} />
-                  <h6 className={Styles.titless}>Banana Smothie Pop</h6>
-                </div>
-              </div>
-              <div className="col-md-4">
-                <div className={Styles.wrappe1}>
-                  <img src={P4} alt="" className={Styles.products} />
-                  <h6 className={Styles.titless}>Cofee Lava Cake</h6>
-                </div>
-              </div>
-              <div className="col-md-4">
-                <div className={Styles.wrappe2}>
-                  <img src={P5} alt="" className={Styles.products} />
-                  <h6 className={Styles.titless}>Sugar Salmon</h6>
-                </div>
-              </div>
-              <div className="col-md-4">
-                <div className={Styles.wrappe3}>
-                  <img src={P6} alt="" className={Styles.products} />
-                  <h6 className={Styles.titles}>Indian Salad</h6>
-                </div>
-              </div>
+              )))}
             </div>
           </div>
         </section>
